Changed counter variable type

+++ ft_strncmp.c
@@ -40,7 +40,7 @@ static int	comp_char(char c1, char c2)
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	c;
+	size_t	c;
 
 	c = 0;
 	while (c < n && s1[c] && s2[c])
